organized code into smaller files
