+++ creator.html
@@ -15,6 +15,6 @@
         <canvas id="canvas"></canvas>
     </div>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/camanjs/4.1.2/caman.full.min.js"></script>
-    <script src="script.js"></script>
+    <script type="module" src="scripts/app.js"></script>
 </body>
 </html>
